% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertImageCode.R
\name{create_valid_file_name}
\alias{create_valid_file_name}
\title{Create a valid, unique file name.}
\usage{
create_valid_file_name(file_path, file_type = ".png")
}
\arguments{
\item{file_path}{Directory path including name of the file}

\item{file_type}{Extension of the image, e.g. .png}
}
\value{
An unique name for the image for the given path
}
\description{
In case the current file name, for the image, is already taken an index will be added to the file name.
}
