% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertImageCode.R
\name{insert_image_code}
\alias{insert_image_code}
\title{Insert the markdown code for the image in the .Rmd file.}
\usage{
insert_image_code()
}
\description{
Given the image in the clipboard, the corresponding markdown code will be inserted in the .Rmd file. The code
will be generated given the code template from the configuration and the generated file path. It will be inserted with
the rstudioapi package.
Unsaved documents as Untitled1.Rmd can not properly be accessed by rstudioapi. The document has to be saved to disk.
If multiple images are copied only one will be copied.
The image can only be inserted in the source editor not in the console nor the terminal.
}
