% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertImageCode.R
\name{save_clipboard_image}
\alias{save_clipboard_image}
\title{Saving the clipboard image to disk}
\usage{
save_clipboard_image(file_name, dir = getwd())
}
\arguments{
\item{file_name}{The image from the clipboard has to be saved to disk. file_name is the name of the image when saving it to disk.}

\item{dir}{The image from the clipboard has to be saved to disk. dir is the directory to save the image in. By default it is the current working directory.}
}
\description{
The image from the clipboard has to be saved to disk. Mac, Linux and Windows are supported. For Windows Powershell is used. For Linux xclip is required.
}
